pyproject.toml
[tool.poetry]
name = "ai-signal"
version = "0.3.0"
description = "Terminal-based AI curator that turns information noise into meaningful signal"
authors = ["Guglielmo Celata <guglielmo.celata@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/guglielmo/ai-signal"
repository = "https://github.com/guglielmo/ai-signal"
documentation = "https://github.com/guglielmo/ai-signal#readme"
keywords = ["ai", "curation", "terminal", "tui", "textual", "content", "aggregator", "filter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Filters",
    "Topic :: Terminals",
    "Topic :: Utilities",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
packages = [
    { include = "aisignal", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
textual = "^0.87.0"
requests = "^2.32.3"
markdown = "^3.7"
typer = "^0.13.1"
rich = "^13.9.4"
pyyaml = "^6.0"
aiohttp = "^3.9.1"
beautifulsoup4 = "^4.12.2"
openai = "^1.3.0"
python-frontmatter = "^1.0.0"
toml = "^0.10.2"
readability-lxml = "^0.8.1"
html2text = "^2024.2.26"
lxml-html-clean = "^0.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
poetry-dynamic-versioning = "^1.4.1"
textual-dev = "^1.7.0"
flake8-pyproject = "^1.2.3"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
venv/
| dist/
| docs/
| htmlcov/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
src_paths = ["src", "tests"]
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
exclude = ".git,__pycache__,build,dist"
per-file-ignores = "__init__.py:F401,storage.py:W291"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aisignal = "aisignal.cli:app"

